[3422][16.0][FIX] mrp_subcontracting_lot_price_unit, add tests (#30)
Before this commit, the case where a product is received once, with a specific lot and a unit price of 1000, the behavior of stock_lot_price_unit should maintain this price when the same product, with the same lot, is received again, even if the new unit price is 2000. This is to respect the existing value of the unit price associated with that particular lot.

However, when the module mrp_subcontracting_lot_price_unit is installed, this behavior changes. The module will not maintain the existing value if the second receipt of the product is not a subcontracted batch. Instead, it will update the existing unit price of the lot with the new value. In this case, it would update the original unit price of 1000 with the new unit price of 2000.